requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
backoff==2.2.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
diskcache==5.6.3
docker-pycreds==0.4.0
emoji==2.14.1
fastapi==0.115.8
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
gql==3.5.0
graphql-core==3.2.6
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.2.3
packaging==24.2
platformdirs==4.3.6
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
sentry-sdk==2.21.0
setproctitle==1.3.4
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.45.3
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uuid_utils==0.10.0
uvicorn==0.34.0
wandb==0.19.6
weave==0.51.33
yarl==1.18.3
